requirements.txt
pygame==2.5.2
pytest==8.0.0
pytest-mock==3.12.0
PyOpenGL==3.1.7
PyOpenGL_accelerate==3.1.7